+++ src/components/Radio/index.js
@@ -1,2 +1,0 @@
-export { default as RadioSkeleton } from "./Radio.Skeleton"
-export { default as Radio } from "./Radio"