+++ src/index.ts
@@ -16,8 +16,6 @@ export { LinkButton } from "./components/LinkButton"
 
 export { Link } from "./components/Link"
 
-export { RadioSkeleton, Radio } from "./components/Radio"
-
 export { FileUpload } from "./components/FileUpload"
 
 export { SidebarNav } from "./components/SidebarNav"
@@ -73,3 +71,4 @@ export * from "./components/Dropdown"
 export * from "./components/Spacer"
 export * from "./components/Navigation"
 export * from "./components/ProgressBar"
+export * from "./components/PlanIndicator"